backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-storages
Pillow
django-filter
python-dotenv
